Remove unnecessary 'localId' field from Comment model
Reason: It was purely a front-end convenience
Solution/replacement: Tracking a counter in component state,
starts at -1 and decrements every new comment. After the db has
created the new comment (and assigns it a positive, unique ID),
then we can use the id in the same way as localID as key identifiers.

Added benefit, easily tracking saved/unsaved comments.

Now you can have multiple comments in 'Edit' mode.
Temporary style decision: unsaved comments show a thick red border.
Once saved and the db returns success, then the border goes away.

See demo here: https://cl.ly/7855670f39b9

+++ src/Grade.tsx
@@ -273,7 +273,7 @@ class Grade extends React.Component<{ match: { params: { subID: typeof Number } 
     }
 
     for (const file of files) {
-      const index = comments[file.id].findIndex((c: IComment) => c.localId === activeCommentId);
+      const index = comments[file.id].findIndex((c: IComment) => c.id === activeCommentId);
       if (index !== -1) {
         comments[file.id][index].rubricComment = rubricComment.id;
         this.setState({ comments });
@@ -310,15 +310,16 @@ class Grade extends React.Component<{ match: { params: { subID: typeof Number } 
     this.setState({ comments });
   };
 
-  public updateComment = (comment: IComment, file: IFile) => {
+  public updateComment = (commentId: number, newComment: IComment, file: IFile) => {
     const { submission, comments } = this.state;
     if (!submission) {
       return;
     }
 
-    const index = comments[file.id].findIndex((c: IComment) => c.localId === comment.localId);
+    const index = comments[file.id].findIndex((comment: IComment) => comment.id === commentId);
 
-    comments[file.id][index] = comment;
+    comments[file.id][index] = newComment;
+
     this.setState({ comments });
   };
 
@@ -330,7 +331,7 @@ class Grade extends React.Component<{ match: { params: { subID: typeof Number } 
       return;
     }
 
-    const index = comments[file.id].findIndex((c: IComment) => c.localId === comment.localId);
+    const index = comments[file.id].findIndex((c: IComment) => c.id === comment.id);
 
     comments[file.id] = [
       ...comments[file.id].slice(0, index),
@@ -344,15 +345,17 @@ class Grade extends React.Component<{ match: { params: { subID: typeof Number } 
     // - Leave as is. If the DELETE fails, it's not that big of a deal. Just annoying.
     // - Keep comment rendered until DELETE completes
     // - Remove comment render, add in a global page loading icon.
-    fetch(`/api/comments/${comment.id}/`, {
-      headers: {
-        Authorization: `JWT ${localStorage.getItem('token')}`,
-      },
-      method: 'DELETE',
-    }).then((res) => {
-      // Returns '204 No Content' on success
-      console.log('res', res);
-    });
+    if (comment.id > 0) {
+      fetch(`/api/comments/${comment.id}/`, {
+        headers: {
+          Authorization: `JWT ${localStorage.getItem('token')}`,
+        },
+        method: 'DELETE',
+      }).then((res) => {
+        // Returns '204 No Content' on success
+        console.log('res', res);
+      });
+    }
   };
 
   public toggleFinalized = () => {
